Implemented suggested changes by code rabbit

+++ test/Altinn.App.Core.Tests/LayoutExpressions/CommonTests/TestFunctions.cs
@@ -288,7 +288,7 @@ public class TestFunctions
                 e.ValueKind switch
                 {
                     JsonValueKind.String => e.GetString(),
-                    JsonValueKind.Number => e.GetDouble(),
+                    JsonValueKind.Number => e.GetDecimal(),
                     JsonValueKind.True => true,
                     JsonValueKind.False => false,
                     JsonValueKind.Null => null,
